% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdViz-package.R
\docType{package}
\name{lipdViz-package}
\alias{lipdViz}
\alias{lipdViz-package}
\title{lipdViz: Visualization for LiPD Data and Ensemble Analyses}
\description{
Plotting and mapping for LiPD (Linked Paleo Data) datasets and time-uncertain ensemble analyses, including age model ensembles, timeseries ribbons and stacks, correlation, regression, spectral, and PCA results, and site maps. Split out of the 'geoChronR' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickmckay/lipdViz}
  \item Report bugs at \url{https://github.com/nickmckay/lipdViz/issues}
}

}
\author{
\strong{Maintainer}: Nicholas McKay \email{nick@nau.edu} (\href{https://orcid.org/0000-0003-3598-5113}{ORCID})

Authors:
\itemize{
  \item Julien Emile-Geay (\href{https://orcid.org/0000-0001-5920-4751}{ORCID})
  \item Deborah Khider (\href{https://orcid.org/0000-0001-7501-8430}{ORCID})
}

}
\keyword{internal}
