% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.corEns}
\alias{plot.corEns}
\title{Plot an ensemble correlation result}
\usage{
\method{plot}{corEns}(x, ...)
}
\arguments{
\item{x}{a corEns result (from ens::corEns)}

\item{...}{passed to plotCorEns}
}
\value{
a ggplot object
}
\description{
Plot an ensemble correlation result
}
