% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.pcaEns}
\alias{plot.pcaEns}
\title{Plot an ensemble PCA result}
\usage{
\method{plot}{pcaEns}(x, TS, ...)
}
\arguments{
\item{x}{a pcaEns result (from ens::pcaEns)}

\item{TS}{the timeseries object used in the pcaEns analysis (required by plotPcaEns)}

\item{...}{passed to plotPcaEns}
}
\value{
a gridExtra ggplot object
}
\description{
Plot an ensemble PCA result
}
