% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.regressEns}
\alias{plot.regressEns}
\title{Plot an ensemble regression result}
\usage{
\method{plot}{regressEns}(x, ...)
}
\arguments{
\item{x}{a regressEns result (from ens::regressEns)}

\item{...}{passed to plotRegressEns}
}
\value{
a list of ggplot objects
}
\description{
Plot an ensemble regression result
}
