% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.spectraEns}
\alias{plot.spectraEns}
\title{Plot an ensemble spectra result}
\usage{
\method{plot}{spectraEns}(x, ...)
}
\arguments{
\item{x}{a spectraEns result (from ens::computeSpectraEns)}

\item{...}{passed to plotSpectraEns}
}
\value{
a ggplot object
}
\description{
Plot an ensemble spectra result
}
